better.

+++ src/state.h
@@ -5,7 +5,7 @@
 
 struct state_t {
     SDL_Texture *texture;
-    float zoom;
+    int zoom;
     SDL_FRect rec;
     bool can_reset;
     struct state_t *next;
